+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ###Fixed
 - Moved protobuf requirement to base.txt
+- Fixed app_id in config (loggers names)
+- Fixed reply to broker option in task_send
 
-###Added
-- Broker Sender
+### Added
+- Added documentation how to visualize workflows
 
 ## [0.1.2] - 2018-06-12